# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "eShop API Documentation",
        "description": "API endpoints for managing and retrieving products, variants, and other eShop resources.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Local development server"
        }
    ],
    "paths": {
        "/api/tokens": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Create API Token",
                "description": "Exchanges user email and password for a plain-text Sanctum API token.",
                "operationId": "9e2abf50332f11ebe83106d55b788354",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password",
                                    "device_name"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "test@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    },
                                    "device_name": {
                                        "type": "string",
                                        "example": "iPhone 15 Pro"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Token created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "1|abcdef123456..."
                                        },
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Test User"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "format": "email",
                                                    "example": "test@example.com"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid inputs or credentials"
                    },
                    "429": {
                        "description": "Too Many Requests (Rate Limit)"
                    }
                }
            }
        },
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get list of active products",
                "description": "Returns a paginated list of active products with their associated variants.",
                "operationId": "d97d1c3727cbac49708b85f7d11f7d58",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response with products and pagination data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "iPhone 15"
                                                    },
                                                    "slug": {
                                                        "type": "string",
                                                        "example": "iphone-15"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "The iPhone 15 features..."
                                                    },
                                                    "summary": {
                                                        "type": "string",
                                                        "example": "Durable design, 48MP..."
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "example": "active"
                                                    },
                                                    "thumbnail": {
                                                        "type": "string",
                                                        "example": "http://..."
                                                    },
                                                    "images": {
                                                        "type": "array",
                                                        "items": {
                                                            "type": "string"
                                                        }
                                                    },
                                                    "options": {
                                                        "type": "array",
                                                        "items": {
                                                            "type": "object"
                                                        }
                                                    },
                                                    "variants": {
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "id": {
                                                                    "type": "integer",
                                                                    "example": 1
                                                                },
                                                                "name": {
                                                                    "type": "string",
                                                                    "example": "iPhone 15 - Black / 128GB"
                                                                },
                                                                "sku": {
                                                                    "type": "string",
                                                                    "example": "IPH15-BLK-128"
                                                                },
                                                                "barcode": {
                                                                    "type": "string",
                                                                    "example": "195949033348"
                                                                },
                                                                "price": {
                                                                    "type": "number",
                                                                    "format": "float",
                                                                    "example": 799
                                                                },
                                                                "compare_at_price": {
                                                                    "type": "number",
                                                                    "format": "float",
                                                                    "example": 899,
                                                                    "nullable": true
                                                                },
                                                                "inventory_quantity": {
                                                                    "type": "integer",
                                                                    "example": 50
                                                                },
                                                                "track_inventory": {
                                                                    "type": "boolean",
                                                                    "example": true
                                                                },
                                                                "continue_selling_out_of_stock": {
                                                                    "type": "boolean",
                                                                    "example": false
                                                                },
                                                                "weight": {
                                                                    "type": "number",
                                                                    "format": "float",
                                                                    "example": 171
                                                                },
                                                                "weight_unit": {
                                                                    "type": "string",
                                                                    "example": "g"
                                                                },
                                                                "dimensions": {
                                                                    "properties": {
                                                                        "width": {
                                                                            "type": "number",
                                                                            "format": "float",
                                                                            "example": 7.16
                                                                        },
                                                                        "height": {
                                                                            "type": "number",
                                                                            "format": "float",
                                                                            "example": 14.76
                                                                        },
                                                                        "depth": {
                                                                            "type": "number",
                                                                            "format": "float",
                                                                            "example": 0.78
                                                                        },
                                                                        "unit": {
                                                                            "type": "string",
                                                                            "example": "cm"
                                                                        }
                                                                    },
                                                                    "type": "object"
                                                                },
                                                                "options": {
                                                                    "type": "object",
                                                                    "example": {
                                                                        "Storage": "128GB",
                                                                        "Color": "Black"
                                                                    }
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/products/import": {
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Bulk import products from CSV files",
                "description": "Uploads multiple CSV files matching the naming pattern `YYYY_MM_DD_timestamp_productName_number.csv` and queues them for processing via RabbitMQ.",
                "operationId": "dd7d5171cdd91a805518de03df641cc4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "files"
                                ],
                                "properties": {
                                    "files[]": {
                                        "description": "List of CSV files to import",
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Files accepted and queued for processing",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "CSV files have been successfully queued for importing."
                                        },
                                        "queued_files": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "example": "2026_06_30_1782860289_product_1.csv"
                                            }
                                        },
                                        "invalid_files": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation failure or all files failed naming format check"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Authentication",
            "description": "Authentication"
        },
        {
            "name": "Products",
            "description": "Products"
        }
    ]
}